git/gitattr/attr_test.go: check third line, too
When writing this test in 58ae7f7f (git/gitattr: introduce package
'gitattr', 2018-07-06), code was introduced to check each line of a
multi-line .gitattributes file.

Since each line was similar, there's a good chance that we did some
over-eager copy/pasting, and wound up checking the same thing twice.

To fix this, let's increment the index of the line that we're testing,
to ensure that all lines are tested correctly.


Former-commit-id: d765b228d221aa797be7d5628d8e1e86d0b7632a