t: make credential helper handle local URLs
The lfstest credential helper provides credentials found by reading a
file on the system.  Currently, it expects that there is a colon in the
URL followed by one or more slashes.  If there is not (say, because
we're using a file:/// or cert:/// URL), it fails to concatenate the
path properly, and instead of looking in the credentials directory,
looks for a different file one directory higher.  This occurs because
filepath.Join strips off a trailing slash if the second argument passed
to it is empty.

This, of course, makes the credential helper fail to work and the tests
hang waiting for input.  Insert an additional slash in the environment
variable and handle the empty string case properly so that credentials
for local paths can be looked up properly.


Former-commit-id: 1790168bd587e3cfb780f76e5bb99beff07317a5