t: generate an encrypted private key for tests
In the future, we'll want to run integration tests that use an encrypted
private key.  To make this as easy as possible, create an encrypted
private key as part of the test setup.  Set up credentials for this key
using the credential helper.

Note that in order to get Git to use the credential helper for the
encrypted private key instead of prompting using OpenSSL, we need to
set GIT_SSL_CERT_PASSWORD_PROTECTED, which makes Git think that the
certificate is password-protected, too.  Therefore, specify credentials
for both the certificate (which isn't really encrypted) and the
encrypted private key.


Former-commit-id: 52f94c2f70ccdc9b4f5d987faf83e6c18e0fbb73