t: make testsuite run under git rebase -x
When making a complex change to Git LFS (or any other project), it is
common to write a multiple patch series.  When doing so, it's ideal to
ensure that each patch is logical and independent and that the tests
pass at each step.  Often this is done using "git rebase -x" to run the
testsuite at each step.

Make the testsuite able to be run under "git rebase -x" by unsetting
environment variables that cause the tests to fail or to operate on the
git-lfs repository itself.  In the Makefile, use a subshell with unset
because env -u is not in POSIX and is supported by neither NetBSD nor
OpenBSD.


Former-commit-id: 46cdbb659ab16db638253116354ed6205c00c6bd