lfsapi: pass stderr through when filling credentials
When git credential invokes a credential helper, it passes through
standard error to the credential helper.  This is helpful because if an
error occurs, it’s printed to the same place other errors go: either the
terminal, or to a log file.

However, Git LFS does not pass standard error through.  We originally
read it git credential to propagate up, but stopped doing so due to a
problem where buffering caused the credential helper to hang.  However,
when we made that change, we unintentionally redirected standard error
to /dev/null.

In addition to being unhelpful to the end user, this also causes
problems in our CI environment when credential failures happen: we end
up with failures that are nearly impossible to troubleshoot because we
lack the full debugging information that would come from standard error
output.  For all of these reasons, wire up git credential's standard
error to os.Stderr.


Former-commit-id: 6df606c19e12e876589cdb504493d0ad4e601154