t: tighten regular expression in push test
In this test, we look for certain lines beginning with "push" by
grepping through the log for lines which contain the word "push".  While
that works now, in the future, the credential helper will start to
prattle to standard error.  In doing so, it will print the name of the
repo, which includes the word "push".  To avoid test failures, tighten
up the regexp to match only at the beginning of the line and only when
there's a space following "push".


Former-commit-id: c37145dc668c234153b949e45f7647f616c1c5b1