Makefile: add a target to automate building Windows assets
Add a target, release-windows, that builds and signs the Windows release
assets.  The target assumes that signtool.exe and iscc.exe are in the
user's PATH.  Since Git for Windows includes unzip, but not zip, use tar
to store the Windows assets for transfer from the Windows machine to the
Unix machine performing the release build.

Additionally, add a target, release-windows-rebuild, to take the tarball
from Windows and repack its contents into the releases zip files. Note
that we cannot do this on Windows because we lack the zip utility as
part of Git Bash.


Former-commit-id: dae16f3806b24638acdecee6bf79ec55e86d5377