commands/command_migrate.go: return all references by remote
We have used the getRemoteRefs since [1] to return a slice of all remote
references, without distinguishing which remote they came from.

The reason for doing so was that we had had modified each reference's
"name" (i.e., the last element of splitting the full reference name by
'/').

This was OK to do, since we modified the reference's name (i.e., the
last component of the fully qualified reference name as read from Git
when splitting on the '/' character) to include the remote name. This
allowed us to call '(*git.Ref).Refspec()', which would theoretically
format us a correctly printed reference name, for all kinds of
references.

This is broken in practice, so to prepare for a future commit that will
fix it, let's return a map of remote name to all references present on
that remote, such that we can format each based on its remote and full
name individually.

[1]: ce89eb35 (commands/command_migrate: teach how to find all remote
     refs, 2017-06-09)


Former-commit-id: 3589a9011dc242869bd35a4ada91b7c335140967