git: introduce constants for index stages
Git can store multiple different versions of a file in the index.
Normally, index stage 0 is the only version provided, but if there's a
conflict, there can be three additional stages.  Since we'll be working
with these stages in a future commit, add constants for them to the git
package.


Former-commit-id: c0a68086e97a127dedd9f0015bec9ef3ab6f4a45