docs/proposals: propose an SSH protocol
There has been a significant amount of interest in a purely SSH-based
protocol for Git LFS.  Add a proposal for an SSH-based protocol based on
Git's pkt-line and protocol v2, while still mapping cleanly onto HTTP
for ease of implementation on the server-side.

Document the basic upload and download protocol, as well as locking.
Coalesce the lock listing and lock verification endpoints in the JSON
API into one that provides both features.

Take care that we allow paths and refs to contain arbitrary byte
sequences, including spaces, by always providing them in an unambiguous
way at the end of a pkt-line.


Former-commit-id: 8563c49498ade84f38e1bc8adf6b1a15e28ab1ba