config: clarify comment on author vs committer
The comment for the CurrentCommitter function used the text "author a
commit".  This is confusing, because the author and committer need not
be the same person.  The comment implied one wanted to produce the
author, whereas the function name implied one wanted to produce the
committer.

In a future version, we'll add a function for the author as well, so
let's standardize on this function being for the committer and update
the comment accordingly by saying "commit a change".


Former-commit-id: 0519ab182fdb2615b151c7454070edc2843a6ece