config: improve lookup of committer name and email
Git provides several different ways to acquire a committer's name and
email.  First, it looks for special GIT_COMMITTER_NAME and
GIT_COMMITTER_EMAIL environment variables.  If these are not set, it
looks in the configuration, and failing that, acquires the email from
the EMAIL environment variable.  This last option is especially common
among users who share dotfiles across machines as it means there is only
one place to modify one's email across all programs.

Read these settings in the order that Git reads them, and add tests that
check that we read them in the order Git does.  Note that unlike Git, we
don't query the system for the username and email address.  This is not
particularly portable across platforms and tends not to contain useful
values.  If there is a compelling use case in the future, we can add
support for it then.


Former-commit-id: 4fd69115c00db49185f2e9c4e5f3925f239cbe9a