vendor: remove extra spew library
We have historically removed the github.com/davecgh/go-spew library,
because we do not depend on it and therefore would not like to carry its
code.

The 'make vendor' recipe contains a couple of $(RM) invocations to
perform that removal, but it appears that some go-spew code has crept
in via 902e855f (go.mod: update gitobj to the latest master,
2018-08-31).

Before updating the latest changes from gitobj, let's run 'make vendor'
to remove this code, so that our subsequent change is cleaner.


Former-commit-id: f51966281735516de2004632b096202c723fd8aa