commands: list explicitly excluded patterns separately
Currently, if a user has disabled Git LFS for a pattern in a
.gitattributes file (such as in a subdirectory) by writing
"-filter=lfs", we render the pattern as a tracked pattern in "git lfs
track", even though those paths are not enabled for Git LFS.

Improve this situation by listing excluded patterns as well as tracked
patterns.  Compute whether a pattern is excluded by looking for an
attribute entry that starts with "-filter" and if so, list the entry as
excluded.  Add tests for both the "-filter" and "-filter=lfs"
situations.


Former-commit-id: 003bce8f8a3b1a7b74b5a6452ce5e6007e583495