CONTRIBUTING.md: no longer require GOROOT, GOPATH, etc
Since the advent of [1], and [2], we use Go 1.11 with modules enabled.
This relieves us of the requirement of having Git LFS checked out within
the appropriate location beneath the caller's $GOPATH.

So, let's remove the guideline, and recommend away from using `go get`,
since this command behaves differently based on whether or not its CWD
is itself a Go repository.

Instead, recommend that we use `git clone`, and do not specify a
location to perform the clone in, since Git LFS can be checked out from
anywhere.

[1]: dfc0f2fc (Merge pull request #3298 from git-lfs/ttaylorr/go-1.11.1,
     2018-10-08)
[2]: 35fe301c (Merge pull request #3208 from git-lfs/ttaylorr/go-mod,
     2018-08-29)


Former-commit-id: fedcef679dbcc065e781c3e62fbbfb9573d01d78