vendor: don't remove necessary dependencies
Since before the Makefile was introduced, we have removed the go-spew
and go-difflib dependencies, since they are not necessary in order to
build or develop Git LFS.

Homebrew running Go 1.11.1 during build, however, requires them. In the
interim before this commit is released, we run `make vendor RM=true` to
refresh the vendor tree without removing these dependencies. In the
future, we'll be able to remove this changed introduced into Homebrew by
not removing the dependencies in the first place.

Since this commit removes the only usage of $(RM), we remove its
definition, too.


Former-commit-id: 1a8965bcd8c5bfceb97c5cfbeb646d1f25c21d06