t/git-lfs-test-server-api: fix reversed conditional
A refactor to this code, 3180a2d0 ("tq: use Rel() function to implement
multiActionSet", 2017-03-06), transformed various lookups such that when
a link, ok pair was returned, a check for ok became a check for link
being non-nil, and a check for !ok became a check for link being nil.
This is consistent with what the actual code does in this case: if
there's no link, a nil is returned for the link, and a nil error is
returned; otherwise, a valid link is returned, possibly with an error.

However, one of the places refactored was accidentally reversed, so we
complained that a link should not exist, and then reported that the
value received was in fact nil (which means that the link does not
exist—exactly what we were hoping for).  Reverse the sense of the check,
so that we're testing what we want and so that this test program passes
when run.


Former-commit-id: f8a659c8551ef7908433873e51a7a4a024bcd043