lfs: don't close pointer scanner on error
On error, we used to attempt to close the pointer scanner to free the
resources that were associated with git cat-file --batch.  However, now
that we're using a gitobj-based scanner, there are no resources of that
kind to free.  Moreover, we only return an error if we return nil for
the scanner object, so attempting to close the scanner leads to a nil
pointer dereference.

Remove the calls to close the scanner when an error occurs.

Signed-off-by: brian m. carlson <bk2204@github.com>

Former-commit-id: 7dae495442111ea09000ee7ebeb7f9c61ba3d66c