git: unify attribute parsing
We currently have two separate pieces of code for parsing attributes:
the attribs.go file in the git package, and the gitattr package.  Unify
them by converting the former to call the latter.  In doing so, move the
line ending splitter from the former package to the latter package and
have the ParseLines call in the latter package return an additional
value which represents the proper line ending value for the file read.


Former-commit-id: 28f96008df463a819b7c17ac9904758edea6d950