git: detect which gitattributes files support macros
Git supports macros in top-level gitattributes files only, not in files
in working tree subdirectories.  When enumerating gitattributes files in
the working tree, determine whether we should read macros from this file
as well.

Note that we always read macros from the system and global files, so
don't touch these code paths for now; we'll update them in a future
commit.


Former-commit-id: bef2aa5db03d4f754a85edc9cc8583c71c14dfcb