tools: introduce a function to look up config files
Several different configuration files used by Git can be located with a
directory specified by the environment variable $XDG_CONFIG_HOME.  Using
this path allows the user to specify a desired place for their
configuration files.  Add a function to look up paths that use this
directory and add tests for it.

Note that we use the slash syntax even on Windows when invoking
ExpandPath as that function doesn't accept backslashes.


Former-commit-id: c2505b7b003dccb033251950c7b759883aabc79b