docs/man: suggest using Git configuration for LFS keys
The documentation suggests using .lfsconfig in the repository root for
configuration keys.  Only a small number of keys can be set in this file
for security reasons, but we don't mention this until the end of the
file, which means that users often don't notice this fact, leading to
confusion when it doesn't work.

Instead, suggest primarily using the Git configuration files for storing
this data, since that will always work.  State that users can use
.lfsconfig, but tell them up front that it is restricted to a small
number of keys, and point them farther down the page where they can find
that information.  Indicate to users that this restriction is for
security reasons to help them better understand the rationale.


Former-commit-id: 9b6b57b2857baa1facb2a0ec68f2f1f073137a2b