commands/uploader: improve error message on missing object
When attempting a push, we need to ensure that we have every object we
want to push.  The object scanner produces pointer objects that contain
the object ID of the object we want to push plus the path at which the
scanner found the object.  Before pushing, we iterate over every object
we want to push and if the object is not found in the LFS object store,
we attempt to read the file that is at the path in question to try to
create the object.

If that attempt fails since the file is no longer present (i.e., it was
present in a previous revision but has since been deleted), we produce
an error message that the file could not be opened without any context,
causing the user to be confused as to why we would have attempted to
open a file that has since been deleted.  Improve this error message by
indicating what the actual problem is (that we cannot find the source
file), a possible step that they can take to solve it (running git lfs
fetch --all), and the error message when attempting to open the file.


Former-commit-id: fea77e1c6f715964d6ebbd8e71374c1896cad9ca