docs/man/git-lfs-migrate: make examples less confusing
When invoking git lfs migrate --no-rewrite, we take a list of files, not
a list of patterns. This differs from other situations, and the examples
in the man page don't make this clear, resulting in confusion on the
part of users. Improve the example by making the example commit message
explain the situation a bit better.


Former-commit-id: 84f8a1405b1a057182ba258d739716048243d293