command/env: ensure we honor lfs.url
When we attempt to look up an endpoint to push to or pull from a remote,
we call the endpoint finder's Endpoint method, which honors lfs.url, not
the RemoteEndpoint method, which does not. However, when enumerating
remotes in git lfs env, we call RemoteEndpoint, which means we produce
the wrong endpoint if the lfs.url option is set.

Update the code to call the correct method, so that we get the correct
results. Since the tests check for this case and print the old
information (which doesn't match what we actually do), update them as
well. Now that we have several tests that produce similar output because
they all use lfs.url, update one of the tests to test more useful cases
by removing the use of lfs.url.


Former-commit-id: 950559b2e86cf427729e3b88a008fad33390f9a8