docs/man/git-lfs-config.5.ronn: document GIT_LFS_SKIP_SMUDGE
Since its inception in [1], the `GIT_LFS_SKIP_SMUDGE` environment
variable has not been documented. This feature is not meant to be
hidden, so let's document how it works in `git-lfs-config.5` in case
users are interested in it.

Let's write this in `git-lfs-config(5)` since this is where environment
variables that aren't tied to a particular command go. We note its
existence in the related manual pages (e.g., `git-lfs-smudge(1)` and
`git-lfs-filter-process(1)`) but refer to the authoritative source in
`git-lfs-config(5)`.

[1]: 8cd51c00 (rename the options to SKIP_SMUDGE and --skip-smudge,
     2015-09-23)


Former-commit-id: 04d5a84ebb98fbcc039dfcf2486b1791d11b8daa