.circleci: don't use 'brew prune'
In the latest release of Homebrew (which Circle CI has upgraded to for
us) the command `brew prune` has been removed.

In order to stop failing our Circle CI builds, let's remove this
invocation.

The recommended alternative is to run 'brew cleanup', but this is
unnecessary: Homebrew 2.0 already does this for us, so repeating it here
is unnecessary.

Former-commit-id: 9cb4c63183e2a2ce1cd80fbae78079dcd1695770