t/t-smudge.sh: remove unnecessary test
In the "smudge with skip" test (i.e., the one that tests using
$GIT_LFS_SKIP_SMUDGE), we unset the variable in the environment, and
then check that it has indeed been removed.

We expect that the 'unset' program will work without issue, so let's
remove the check that ensures it so.

As an additional benefit (and the reason for this change in the first
place) we allow this test to continue to pass even when the latest
commit message mentions the word "GIT_LFS_SKIP_SMUDGE".

On some CI systems, the commit message from HEAD is placed in the
environment variable (e.g., APPVEYOR_REPO_COMMIT_MESSAGE_EXTENDED) and
hence causes this test to fail.


Former-commit-id: 32f2490258aea55b877ae93c33fc8ffaf5a3fb34