tq: avoid nil pointer dereference on unexpected failure
When we get an error making an API request, we check to see if the
response is a 429, and if so, we attempt to back off and retry later.
However, we failed to check if we actually had a response; if the
response was missing (say, because we had a network issue), then we
would dereference a nil pointer.

Instead, let's explicitly check for a non-nil response before querying
it for a status code.

Co-authored-by: Taylor Blau <me@ttaylorr.com>

Former-commit-id: 4775f9c19c239b4bca2f7a6759c089a7d84b3e90