Switch from manually running go generate to using dh-golang to run it
This fixes #2013.  For some reason I could only reproduce the issue from
within the docker builder, not when running outside of docker.
Regardless, this appears to fix the problem and is a generally cleaner
solution anyway that will continue to work if any other generate
commands are added to git-lfs.


Former-commit-id: 8b90eff16e65622553e387eecff9a16cc768083a