commands: sanity-check lock paths
When creating a lock, the file to be locked should always be relative to
the root of the repository, and not contain any ".." components.
Previously, such a ".." component could occur on Windows when using a
junction because the repository root we received from Git was not
properly canonicalized.

Now that that issue has been fixed, it should no longer be possible to
create a lock with a ".." component. However, as a safety feature, let's
check for this anyway and report it as an error if it occurs.


Former-commit-id: 4a9898269d84b5452d5fb4118dcde8ecb3ab6de3