Log test server stderr to log
When there's a problem with the test server, it's helpful to see the log
output to help troubleshoot the issue. However, when the Stderr member
of a command struct is not set, the standard error gets redirected to
/dev/null, which isn't really what we want.

Since we already have a log set up for logging standard output, send
standard error there as well.


Former-commit-id: 420217c2a2ce0445fe4d444afe3e0a6b55d8c2d1