tasklog: detect Cygwin-style pipe TTYs as TTYs as well
Some Unix emulation layers for Windows, such as Cygwin and MSYS, use
pipes to create their Unix-style TTYs. These TTYs are not detected as
such by Windows, which sees them as pipes. However, our isatty module
has a way to detect them, and they will function just like a TTY for our
purposes, so check for them as well and handle them as TTYs if present.


Former-commit-id: 705f4f032596887ac7a306e235ee1772aeb779ea