locking: new "locks --verify" option
"locks --verify" verifies the server-side lock owner and denotes own
locks in the output by 'O'.

Internally, the /locks/verify API is used to have the server reporting
locks categorized by "ours" and "theirs". Note that this categorization
must be performed on the server because only the server knows the
mapping between login and lock owner name.

This basically resolves issue #3252 but does not yet address the
combination of "--verify" and "--cached".


Former-commit-id: ffe6220b62ec100dec4b62afd77606a37c094ddb