config: allow vendors to customize the version info
It's common for vendors (such as Linux distros) to want to customize the
User-Agent string and version information for software they provide.
This is valuable for vendors, but it also allows troubleshooting
problems with custom patches that may have been applied by the vendor.
Allow vendors to set a VENDOR makefile flag to set the vendor field for
the User-Agent and version information.


Former-commit-id: a8c3d8f070e89a2b4fc5782519e87c17c68bb0c4