tq: ensure we pass the correct Accept header in verify requests
When we perform a verify request, we need to specify a proper Accept
header along with the Content-Type header. Do so.

Hoist the setting of the headers before the code which copies headers
from the remote side, so that the remote side can override it with a
suitable header if necessary. This is required for compatibility with
some existing servers.


Former-commit-id: 61bc880d9ede1335e408a21a992ed3ce1e6ea7c4