CONTRIBUTING: encourage a rationale in the pull request
Sometimes we get pull requests that have no description or rationale.
This makes it hard for maintainers to evaluate why the pull request is
needed and what problems it is intended to solve. Encourage users to
provide some explanation in the pull request and hinting that they may
use part (or all) of their good commit message to accomplish this goal.


Former-commit-id: ff57c2e6923cc3698ccd2c44579037487b06ca8c