docs: be more precise about what timestamps we accept
In a couple of places in our documentation, we specified that a
timestamp should be in ISO 8601 format. However, ISO 8601 specifies a
large number of variants, and Go only parses a subset of those. For
example, using a timestamp without the colon in the time zone offset
will cause Git LFS to fail.

Be explicit about what we want as a timestamp, which is an uppercase
RFC 3339-formatted timestamp with second precision. That's what Go
parses, and that's what we expect. It is, in addition, a profile of ISO
8601, so other tools that can parse the entirety of ISO 8601 will still
accept these date formats.


Former-commit-id: c5840d5df9a01a3f625cbe648dc877d1e119872f