githistory: rewrite tags pointing to tags
When rewriting history, we've ignored tags pointing to other tags. This
leads to problems for folks that are trying to shrink the size of their
repository, since they may still have tags that point to the old
history. Add additional handling for tags pointing to tags, recursively
rewriting the inner tags until we reach a commit.


Former-commit-id: ff410fa76120a312402a9e8c0793166e997516b1