git/githistory/rewriter.go: uncache with mode
When we rewrite a blob, we cache certain data based on the file name and
the old object ID. However, we don't consider the mode, so if we have
two identical blobs with the same relative name but different modes,
we'll write the mode we first saw into all subsequent blobs.

Fix this by copying the mode across from the existing blob when we
rewrite the new blob.

Co-authored-by: brian m. carlson <bk2204@github.com>

Former-commit-id: a470c39587bb9a5ba9f583dccbd4cf95b7266b4f