debian/rules: update for Go 1.12
dh_golang wants to turn off the module cache because attempting to use
it breaks sbuild, a component of Debian's package build system. However,
as of Go 1.12, the package build cache is required. Create a temporary
directory for the cache and override the environment variable so that we
can build with Go 1.12. Pass the environment variables to the dh command
directly, since export only affects submakes, not individual commands.

In addition, stop copying the pkg directory into the archive, because it
is no longer produced and attempting to copy nonexisting files causes
the build to fail.


Former-commit-id: 20b25b975f57025f866743701fa407a1cab30439