circleci: remove blank lines from config
CircleCI refuses to accept a configuration file with blank lines,
claiming that it is a version 1.0 configuration file, despite the
command-line tool accepting it as a valid version 2.0 file. To please it
and restore our CI configuration to functionality, remove the blank
lines.

Note that the YAML specification does permit blank lines, even if
CircleCI does not.


Former-commit-id: f7b1df79be76ebdd29dc1320f01f26a208d3618e