track: use default line endings for core.autocrlf=input
We currently use core.autocrlf to guess what line endings a user might
want to have for the .gitattributes file if it lacks line endings
already. When the user has specified core.autocrlf=input, they have
specified that they don't want line endings to be changed. This isn't
explicitly an indication that they want to use CRLF line endings,
especially if they are on a non-Windows system.

Therefore, if we need to guess what line endings to use with
core.autocrlf=input, guess the system line endings, not CRLF.


Former-commit-id: 226f5bb092125cab760759863219d125c6432592