ssh: redirect stdin from /dev/null
When using SSH cloning with a ControlMaster socket and a password on a
Unix system, we're unable to read from standard input and consequently
the SSH process fails. The redirect from standard input was intended to
fix an issue on Windows that is due to /dev/tty disappearing, but it
doesn't appear to work and causes issues on Unix systems. Revert this
change so we can explore other options.

This reverts commit cbea0b915881f3d1ed68b4dbfedef87cf0d15435 [formerly 59ed04855afd7e8f1bb126eb6f6a5c5297f59253].


Former-commit-id: 296311634cb832604cc420f99f689f7b828cb36f