Remove unused pty code
The code which allocated and used a pty on Unix systems was unused, but
it did bring in a dependency on the github.com/kr/pty module. This
module is unmaintained and doesn't compile with gccgo. Since we no
longer need the pty code or this module, remove both of them.


Former-commit-id: 65c6bcf7fdb9c0f73bfbc3e8560f71f04d7b1d91