smudge: ensure custom transfer agent work
When we use the smudge call (as opposed to filter-process), we want to
be able to look up a custom transfer agent based on the URL we're
fetching from. In order to do so, pass the remote we're using in order
to find the right URL and therefore the right custom transfer agent.


Former-commit-id: 183fd6df642d9500004cce39f8480acdcf3664dc