tools: detect MINGW as Cygwin
There are several cases in which we can pass in Unix-style paths in the
Git Bash environment, which is MINGW64-based, including in file URLs. To
make such paths work, let's accept such URLs with the Unix-style paths
and then translate them to Windows-style paths automatically. Note that
passing Windows-style paths to "cygpath -W", which is what we use in
this case, works fine, so there's no regression in the Git Bash case.


Former-commit-id: baa79b1684841e4a0f350c3f1026f78cabaf11dc